requirements.txt
# Core LangChain packages - use compatible versions
langchain-core>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<0.4.0
langchain>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-elasticsearch>=0.4.0,<1.0.0
langgraph>=0.2.0,<1.0.0

# Elasticsearch
elasticsearch==8.17.2

# OpenAI
openai>=1.0.0,<2.0.0

# PDF processing
pypdf>=3.0.0,<4.0.0

# Display and utilities
rich>=13.0.0,<14.0.0
python-dotenv>=1.0.0,<2.0.0

# ML/AI packages (optional - can be installed separately if causing issues)
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<3.0.0
nltk>=3.8.0,<4.0.0
langchain-tavily